move getOMSdata and SetVariable as external functions
